+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * ensuring test coverage is always captured in CI
   [#210]((https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/210)
 
+### Changed
+
+* Suppressing CSW dependency warnings
+  [#224](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/224)
+
 ## [0.4.3] - 2022-03-08
 
 ### Changed